Fixed all source-code and project warnings to give clean compiles. Tracked down crash bug in portal generation.

+++ trunk/ darkbasicpro --username LeeBamberTGC/Dark Basic Pro SDK/Shared/Q3BSP/Shaders.cpp
@@ -1211,7 +1211,7 @@ BOOL Q3ARenderTextureFaces ( int stex )
 				if ( s->stage [ st ].ani_freq > 0 )
 				{
 					int texs = s->stage [ st ].tex_count;
-					int pic  = s->stage [ st ].ani_freq * texs * Timer ( ) / 10;
+					int pic  = (int) (s->stage [ st ].ani_freq * texs * Timer ( ) / 10);
 					
 					pic %= texs;
 					a = pic;
